added wild route and added home route
