+++ public/css/homepage/home.css
@@ -1,0 +1,331 @@
+@import url("https://fonts.googleapis.com/css2?family=Baloo+Bhai+2&display=swap");
+html {
+  scroll-behavior: smooth;
+}
+
+* {
+  box-sizing: border-box;
+  top: 0;
+  font-family: "Baloo Bhai 2", cursive;
+}
+
+body {
+  background: white;
+  color: #f9f9f9;
+  font-family: Avenir-Roman, sans-serif;
+  margin: 0;
+  padding: 0;
+}
+
+a {
+  color: #f9f9f9;
+  text-decoration: none;
+}
+.login {
+  background-color: rgba(0, 0, 0, 0.6);
+  padding: 8px 16px;
+  text-transform: uppercase;
+  letter-spacing: 1.5px;
+  border: 1px solid #f9f9f9;
+  border-radius: 4px;
+  transition: all 0.2s ease-out;
+}
+.login:hover {
+  background-color: #f9f9f9;
+  color: #000;
+  border-color: transparent;
+  cursor: pointer;
+}
+.box {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  height: 70px;
+  background-color: #0c0f1a;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 0 36px;
+  /* letter-spacing: 16px; */
+  z-index: 3;
+}
+.logo {
+  padding: 0;
+  width: 45px;
+  margin-top: 4px;
+  max-height: 70px;
+  display: inline-block;
+}
+.logo img {
+  display: flex;
+  width: 100%;
+}
+.logo span {
+  margin-left: -20px;
+  font-size: 30px;
+}
+.container {
+  align-items: center;
+  display: flex;
+  flex-flow: row nowrap;
+  height: 100%;
+  justify-content: flex-end;
+  margin: 0px;
+  padding: 0px;
+  position: relative;
+  margin-right: auto;
+  margin-left: 25px;
+}
+
+.container a {
+  display: flex;
+  align-items: center;
+  padding: 0 12px;
+  text-decoration: none;
+}
+.container span {
+  color: rgb(249, 249, 249);
+  font-size: 18px;
+  letter-spacing: 1.42px;
+  line-height: 1.08;
+  padding: 2px 20px;
+  white-space: nowrap;
+  position: relative;
+}
+.container span::before {
+  background-color: rgb(249, 249, 249);
+  border-radius: 0px 0px 4px 4px;
+  bottom: -6px;
+  content: "";
+  height: 2px;
+  left: 0px;
+  opacity: 0;
+  position: absolute;
+  right: 0px;
+  transform-origin: left center;
+  transform: scaleX(0);
+  transition: all 250ms cubic-bezier(0.25, 0.46, 0.45, 0.94) 0s;
+  visibility: hidden;
+  width: auto;
+}
+.container span:hover::before {
+  transform: scaleX(1);
+  visibility: visible;
+  opacity: 1 !important;
+}
+
+.info {
+  height: 100vh;
+  color: #f9f9f9;
+  font-family: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+.info::before {
+  content: "";
+  background: url("/images/back.jpg") center center / cover no-repeat fixed;
+  position: absolute;
+  top: 0px;
+  left: 0px;
+  height: 100vh;
+  width: 100%;
+  z-index: -1;
+  /* opacity: 0.7; */
+}
+.write {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+  color: #fafafa;
+}
+.write h1 {
+  color: rgb(240, 148, 43);
+  background: linear-gradient(180deg, black, grey);
+  box-shadow: 5px 8px 20px 5px rgba(0, 0, 0, 1);
+  border-radius: 10px;
+  font-size: 100px;
+}
+.btn {
+  font-weight: bold;
+  font-size: 34px;
+  border-radius: 16px;
+  color: white;
+  background-color: rgb(254, 64, 91);
+  padding: 0px 50px;
+  cursor: pointer;
+  transition: all 0.1s ease-in-out;
+}
+.btn:hover {
+  background-color: rgb(229, 74, 97);
+  transition: all 0.1s ease-in-out;
+  scale: 1.2;
+}
+
+.review {
+  color: black;
+  margin-top: 20px;
+  padding: 0 40px;
+}
+.box2 {
+  display: flex;
+}
+.first {
+  flex: 0.33;
+  border: 2px solid lightgrey;
+  border-radius: 10px;
+  margin-right: 10px;
+  padding: 0 40px;
+  box-shadow: 0px 8px 20px 0px rgba(0, 0, 0, 0.15);
+}
+.first:hover{
+  scale: 1.01;
+  box-shadow: 0px 8px 20px 0px rgba(0, 0, 0, 0.3);
+}
+.first h1 {
+  font-size: 17px;
+  border-bottom: 2px solid lightgrey;
+}
+.first p {
+  color: grey;
+}
+
+.legalize {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  width: 80%;
+  margin-top: 20px;
+  padding: 0 10px;
+}
+.column h1 {
+  font-size: 20px;
+}
+
+.bottom_info {
+  display: flex;
+  flex-direction: column;
+}
+.bottom_info span {
+  color: grey;
+  font-size: 18px;
+}
+.bottom_info span:hover {
+  color: rgb(218, 24, 47);
+  font-size: 18px;
+  cursor: pointer;
+}
+.bottom_info span:first-child {
+  margin-top: -10px;
+}
+.row {
+  display: flex;
+  flex-direction: row;
+  align-items: baseline !important;
+  justify-content: center !important;
+}
+.row i {
+  margin-right: 10px;
+}
+
+.app {
+  border-top: 2px solid grey;
+  border-bottom: 2px solid grey;
+}
+.app_info {
+  display: flex;
+  align-items: center;
+}
+.app_info h1 {
+  margin-right: 20px;
+}
+.app_info img {
+  margin-right: 20px;
+}
+.app_img {
+  max-width: 100%;
+  height: 48px;
+  object-fit: contain;
+  cursor: pointer;
+}
+.app_img2 {
+  max-width: 100%;
+  height: 70px;
+  object-fit: contain;
+  cursor: pointer;
+}
+.app_information h5 {
+  color: grey;
+}
+
+.footer {
+  margin-top: 20px;
+  margin-bottom: 40px;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  flex-wrap: wrap;
+}
+.footer_info span {
+  color: rgb(86, 84, 84);
+}
+.footer_info span:hover {
+  color: rgb(218, 24, 47);
+  cursor: pointer;
+}
+
+.footer_right {
+  color: rgb(86, 84, 84);
+}
+
+@media screen and (max-width: 850px) {
+  .write h1 {
+    font-size: 60px;
+  }
+}
+@media screen and (max-width: 768px) {
+  .hide {
+    display: none !important;
+  }
+  .first {
+    flex: 1;
+  }
+  .app_info h1 {
+    font-size: 18px;
+  }
+  .app_img {
+    height: 30px;
+  }
+  .app_img2 {
+    height: 40px;
+  }
+  .btn {
+    font-size: 18px;
+  }
+}
+@media screen and (max-width: 600px) {
+  .left img {
+    align-items: center;
+    justify-content: center;
+    display: flex;
+  }
+  .info {
+    padding: 0px calc(3.5vw + 5px);
+  }
+  .write h1 {
+    font-size: 30px;
+  }
+  .first p {
+    font-size: 10px;
+  }
+  .first h1 {
+    font-size: 14px;
+  }
+}
+@media screen and (max-width: 400px) {
+  .app_info h1 {
+    font-size: 10px;
+  }
+}
