+++ routes/auth/login.js
@@ -6,6 +6,7 @@ const isAuth = require("./isauth");
 const cookieParser = require("cookie-parser");
 const bodyParser = require("body-parser");
 const UserDetail = require("../../src/models/userDetails");
+const User = require("../../src/models/userauth");
 
 const router = express.Router();
 
@@ -20,7 +21,7 @@ router.use(cookieParser());
 
 router.get("/login", async (req, res) => {
     try {
-        const user = await isAuth(req);
+        // const user = await isAuth(req);
         //TODO : uncomment before uploading it
         // if (user) {
         //     res.status(200).redirect("/test2");
@@ -39,6 +40,9 @@ router.get("/login", async (req, res) => {
 router.post("/register", async (req, res) => {
     try {
         const { username, email, password, confirmPassword } = req.body;
+        const data = await User.findOne({email : email});
+        if(!data)
+        {
         if (password === confirmPassword) {
             const user = await createUser(username, email, password);
             const token = await createToken(user._id);
@@ -52,6 +56,10 @@ router.post("/register", async (req, res) => {
             console.log("password didn't matched");
             res.send("password didn't matched");
         }
+    }else{
+        console.log("user already have a account");
+        res.redirect("/auth/login");
+    }
 
     } catch (error) {
         console.log(error);
@@ -67,7 +75,7 @@ router.post("/login", async (req, res) => {
         if (result.status) {
             const data = await UserDetail.findOne({ _id: result.userid }).select({ moreDetail: 1 });
             if (data && data.moreDetail) {
-                res.status(result.statuscode).redirect("/test2");
+                res.status(result.statuscode).redirect("/viewprofile");
             }
             else {
                 res.status(result.statuscode).redirect("/buildprofile/userdetails");
@@ -85,5 +93,24 @@ router.post("/login", async (req, res) => {
     }
 
 });
+
+router.get("/logout", async (req, res) => {
+    try {
+        const user = await isAuth(req);
+        if(user){
+            res.clearCookie("jwt");
+            res.redirect("/auth/login");
+        }
+        else{
+
+        }
+
+    } catch (error) {
+        console.log(error);
+        res.send(error);
+    }
+});
+
+
 
 module.exports = router;
