+++ views/error404/error.ejs
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html lang="en" dir="ltr">
+  <head>
+    <meta charset="utf-8">
+    <title>CUPIT - error 404</title>
+
+    <link rel="preconnect" href="https://fonts.googleapis.com">
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+    <link href="https://fonts.googleapis.com/css2?family=Lobster&family=Patrick+Hand&display=swap" rel="stylesheet">
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
+    <link rel="stylesheet" href="/css/error404/errorpage.css">
+  </head>
+  <body>
+    <div class="card" style="width: 25rem;">
+  <img class="card-img-top" src="https://media1.giphy.com/media/l378yQjDMER6TR8Q0/200w.webp?cid=ecf05e47lsog53td3s3fidsnmo859662fxvsajvq38acl6qv&rid=200w.webp&ct=s" alt="Card image cap">
+  <div class="card-body">
+    <h5 class="card-title">404 PAGE NOT FOUND</h5>
+    <!-- <p class="card-text">Wrong Password Or Username</p> -->
+    <a href="/" class="btn btn-primary">Back To Homepage</a>
+  </div>
+</div>
+  </body>
+</html>
