+++ views/homepage/home.ejs
@@ -1,0 +1,167 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>CUPIT - HOME</title>
+    <link rel="stylesheet" href="/css/homepage/home.css" />
+    <script src="https://kit.fontawesome.com/9542957272.js" crossorigin="anonymous"></script>
+</head>
+
+<body>
+    <div class="box">
+        <div class="logo">
+            <!-- <img src="/images/logo1.png" alt="Logo" /> -->
+            <a href="/">
+                <span>CUPIT</span>
+            </a>
+        </div>
+        <div class="container">
+            <a href="#">
+                <span>Projects</span>
+            </a>
+            <a href="#">
+                <span>Learn</span>
+            </a>
+            <a href="#">
+                <span>Safety</span>
+            </a>
+            <a href="#">
+                <span>Support</span>
+            </a>
+        </div>
+        <a href="/auth/login">
+            <div class="login">Login</div>
+        </a>
+    </div>
+    <section class="info">
+        <div class="write">
+            <h1>Do Something Epic.</h1>
+            <a href="/auth/login">
+                <button class="btn">Create Account</button>
+            </a>
+        </div>
+    </section>
+    <div class="review">
+        <div class="box2">
+            <div class="first">
+                <h1>Sambhav Sharma</h1>
+                <p>
+                    Thanks to Cupit I have found the love of my life and we are to be
+                    married. After going on a few dates and having a few fun nights I
+                    came across Miranda. After reading her profile I couldn’t resist
+                    swiping right after reading her final sentence... ‘Looking for my
+                    super babe for life.’ After talking for about a week we went out on
+                    our first date and I knew there was something special about her!
+                </p>
+            </div>
+            <div class="first hide">
+                <h1>Sarthak Patel</h1>
+                <p>
+                    For all the single people out there, especially introverted ones
+                    like us: do not be afraid to travel outside of your comfort zone.
+                    That’s where you’ll make a genuine connection. Cupit brought us
+                    together and for that, I am forever grateful. ❤
+                </p>
+            </div>
+            <div class="first hide">
+                <h1>Sambhav Jain</h1>
+                <p>
+                    We were at a crossroads and so after a few months of casually
+                    dating, we decided to end things. I was in love, but thought this
+                    love story wasn't meant to last. Fast forward 9 months, and we had
+                    become inseparable.
+                </p>
+            </div>
+        </div>
+        <div class="legalize">
+            <div class="column">
+                <div class="top_info">
+                    <h1>LEGAL</h1>
+                </div>
+                <div class="bottom_info">
+                    <span>Privacy</span>
+                    <span>Terms</span>
+                    <span>Cookie Policy</span>
+                    <span>Intellectual Property</span>
+                </div>
+            </div>
+            <div class="column">
+                <div class="top_info">
+                    <h1>CAREERS</h1>
+                </div>
+                <div class="bottom_info">
+                    <span>Careers Portal</span>
+                    <span>Tech Blog</span>
+                </div>
+            </div>
+            <div class="column">
+                <div class="top_info">
+                    <h1>Social</h1>
+                </div>
+                <div class="bottom_info row">
+                    <span><i class="fa-brands fa-twitter"></i></span>
+                    <span><i class="fa-brands fa-facebook"></i></span>
+                    <span><i class="fa-brands fa-youtube"></i></span>
+                    <span><i class="fa-brands fa-instagram"></i></span>
+                </div>
+            </div>
+            <div class="column">
+                <div class="top_info">
+                    <h1>HELP</h1>
+                </div>
+                <div class="bottom_info">
+                    <span>FAQ</span>
+                    <span>Destinations</span>
+                    <span>Press Room</span>
+                    <span>Contact</span>
+                    <span>Promo Code</span>
+                </div>
+            </div>
+        </div>
+        <div class="app">
+            <div class="app_info">
+                <h1>GET THE APP!</h1>
+                <img class="app_img" src="/images/app.webp" alt="" />
+                <img class="app_img2" src="/images/google.webp" alt="" />
+            </div>
+            <div class="app_information">
+                <h5>
+                    Single people, listen up: If you’re looking for love, want to start
+                    dating, or just keep it casual, you need to be on Cupit. With over
+                    55 billion matches made, it’s the place to be to meet your next best
+                    match. Let’s be real, the dating landscape looks very different
+                    today, as most people are meeting online. With Cupit, the world’s
+                    most popular free dating app, you have millions of other single
+                    people at your fingertips and they’re all ready to meet someone like
+                    you. Whether you’re straight or in the LGBTQIA community, Cupit’s
+                    here to bring you all the sparks. <br />
+                    <br />
+                    There really is something for everyone on Cupit. Want to get into a
+                    relationship? You got it. Trying to find some new friends? Say no
+                    more. New kid on campus and looking to make the most of your college
+                    experience? Cupit U’s got you covered. Cupit isn’t your average
+                    dating site — it’s the most diverse dating app, where adults of all
+                    backgrounds and experiences are invited to make connections,
+                    memories, and everything in between.
+                </h5>
+            </div>
+        </div>
+        <div class="footer">
+            <div class="footer_info">
+                <span>FAQ &nbsp&nbsp</span>/
+                <span>&nbsp&nbspSafety Tips&nbsp&nbsp</span>/
+                <span>&nbsp&nbspTerms&nbsp&nbsp</span>/
+                <span>&nbsp&nbspCookie Policy&nbsp&nbsp</span>/
+                <span>&nbsp&nbspPrivacy Settings&nbsp&nbsp</span>
+            </div>
+            <div class="footer_right">
+                <span>© 2022 Match Group, CUPIT, All Rights Reserved.</span>
+            </div>
+        </div>
+    </div>
+</body>
+
+</html>
