+++ views/login_registeration/profileverification.ejs
@@ -69,7 +69,7 @@
             console.log(json.status);
             if(json.status === 200)
             {
-              window.location.replace("http://localhost:3000/test2");
+              window.location.replace("http://localhost:3000/viewprofile");
             }
           })
           .catch((err) => console.error(err));
