+++ views/uploadprofile/uploadProfile.ejs
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
+    <title>CUPIT</title>
 </head>
 <body>
     <form action="/buildprofile/uploadprofile" method="post" enctype="multipart/form-data">
