+++ views/viewprofile/mainpage.ejs
@@ -5,7 +5,7 @@
   <meta charset="UTF-8" />
   <meta http-equiv="X-UA-Compatible" content="IE=edge" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>Cupit</title>
+  <title>CUPIT</title>
   <link rel="stylesheet" href="css/viewprofile/usersprofileview.css" />
   <script src="https://kit.fontawesome.com/9542957272.js" crossorigin="anonymous"></script>
 </head>
